+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## next
+## 2.1.0 (December, 29, 2019)
 
 ### driver